+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [UNRELEASED] (NAME) - YYYY-MM-DD
 ### Added
-- TBD (Added a new feature)
+- Added support for rfc2622 ^+, ^-, ^n and ^n-m prefix operators
 
 ### Changed
 - TBD (Changed existing functionality)